+++ hax-types/src/diagnostics/mod.rs
@@ -15,7 +15,6 @@ pub struct Diagnostics {
 
 impl std::fmt::Display for Diagnostics {
     fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
-        write!(f, "({}) ", self.context)?;
         match &self.kind {
             Kind::Unimplemented { issue_id, details } => write!(
                 f,
@@ -31,7 +30,7 @@ impl std::fmt::Display for Diagnostics {
             ),
             Kind::UnsupportedMacro { id } => write!(
                 f,
-                "The unexpanded macro {} it is not supported by this backend. Please verify the option you passed the {} (or {}) option.",
+                "The unexpanded macro {} it is not supported by this backend.\nPlease verify the option you passed the {} (or {}) option.",
                 id.bold(),
                 "--inline-macro-call".bold(), "-i".bold()
             ),
@@ -49,10 +48,9 @@ impl std::fmt::Display for Diagnostics {
             ),
             Kind::ExpectedMutRef => write!(
                 f,
-                "At this position, Hax was expecting an expression of the shape `&mut _`. Hax forbids `f(x)` (where `f` expects a mutable reference as input) when `x` is not a {}{} or when it is a dereference expression.
+                "At this position, Hax was expecting an expression of the shape `&mut _`.\nHax forbids `f(x)` (where `f` expects a mutable reference as input) when `x` is not a {}{} or when it is a dereference expression.
 
-{}
-",
+{}",
                 "place expression".bold(),
                 "[1]".bright_black(),
                 "[1]: https://doc.rust-lang.org/reference/expressions.html#place-expressions-and-value-expressions"
@@ -62,15 +60,53 @@ impl std::fmt::Display for Diagnostics {
                 "The bindings {:?} cannot be mutated here: they don't belong to the closure scope, and this is not allowed.",
                 bindings
             ),
-            Kind::ArbitraryLHS => write!(f, "Assignation of an arbitrary left-hand side is not supported. `lhs = e` is fine only when `lhs` is a combination of local identifiers, field accessors and index accessors."),
+            Kind::ArbitraryLHS => write!(f, "Assignation of an arbitrary left-hand side is not supported.\n`lhs = e` is fine only when `lhs` is a combination of local identifiers, field accessors and index accessors."),
 
             Kind::AttributeRejected {reason} => write!(f, "Here, this attribute cannot be used: {reason}."),
 
-            Kind::NonTrivialAndMutFnInput => write!(f, "The support in hax of function with one or more inputs of type `&mut _` is limited. Onlu trivial patterns are allowed there: `fn f(x: &mut (T, U)) ...` is allowed while `f((x, y): &mut (T, U))` is rejected."),
+            Kind::NonTrivialAndMutFnInput => write!(f, "The support in hax of function with one or more inputs of type `&mut _` is limited.\nOnly trivial patterns are allowed there: `fn f(x: &mut (T, U)) ...` is allowed while `f((x, y): &mut (T, U))` is rejected."),
 
             Kind::FStarParseError { fstar_snippet, details: _ } => write!(f, "The following code snippet could not be parsed as valid F*:\n```\n{fstar_snippet}\n```"),
 
             _ => write!(f, "{:?}", self.kind),
+        }?;
+        write!(f, "\n\n")?;
+        if let Some(issue) = self.kind.issue_number() {
+            write!(
+                f,
+                "This is discussed in issue https://github.com/hacspec/hax/issues/{issue}.\nPlease upvote or comment this issue if you see this error message."
+            )?;
+        }
+        write!(
+            f,
+            "{}",
+            format!(
+                "\nNote: the error was labeled with context `{}`.\n",
+                self.context
+            )
+            .bright_black()
+        )?;
+        Ok(())
+    }
+}
+
+impl Kind {
+    fn issue_number(&self) -> Option<u32> {
+        match self {
+            Kind::UnsafeBlock => None,
+            Kind::Unimplemented { issue_id, .. } => issue_id.clone(),
+            Kind::AssertionFailure { .. } => None,
+            Kind::UnallowedMutRef => Some(420),
+            Kind::UnsupportedMacro { .. } => None,
+            Kind::ErrorParsingMacroInvocation { .. } => None,
+            Kind::ClosureMutatesParentBindings { .. } => Some(1060),
+            Kind::ArbitraryLHS => None,
+            Kind::ExplicitRejection { .. } => None,
+            Kind::UnsupportedTupleSize { .. } => None,
+            Kind::ExpectedMutRef => Some(420),
+            Kind::NonTrivialAndMutFnInput => Some(1405),
+            Kind::AttributeRejected { .. } => None,
+            Kind::FStarParseError { .. } => todo!(),
         }
     }
 }